appear to have demand routing working, but lsp testing is failing a lot
