+++ prototypes/igp_routing_topology.csv
@@ -1,0 +1,34 @@
+INTERFACES_TABLE
+node_object_name	remote_node_object_name	name	cost	capacity    circuit_id
+A	B	A-to-B	20	125 1
+B	A	B-to-A	20	125 1
+B	D	B-to-D	20	125 2
+D	B	D-to-B	20	125 2
+A	C	A-to-C	30	150 3
+C	A	C-to-A	30	150 3
+D	C	D-to-C	30	150 4
+C	D	C-to-D	30	150 4
+A	E	A-to-E	10	300 5
+E	A	E-to-A	10	300 5
+D	F	D-to-F	10	300 6
+F	D	F-to-D	10	300 6
+A	D	A-to-D	40	20  7
+D	A	D-to-A	40	20  7
+B	G	B-to-G	10	100 8
+G	B	G-to-B	10	100 8
+G	D	G-to-D	10	100 9
+D	G	D-to-G	10	100 9
+
+NODES_TABLE
+name	lon	lat
+A	50	0
+B	0	50
+C	0	50
+D 50	0
+E	75	0
+F 75	0
+G	30	30
+
+DEMANDS_TABLE
+source	dest	traffic	name
+A	F	40	dmd_a_f_1
