+++ prototypes/multidigraph_topology_2.csv
@@ -1,0 +1,27 @@
+INTERFACES_TABLE
+node_object_name    remote_node_object_name name    cost    capacity    circuit_id
+A	B	A-to-B	4	100 1
+B	A	B-to-A	4	100 1
+A	B	A-to-B_2	4	100 2
+B	A	B-to-A_2	4	100 2
+A	B	A-to-B_2	4	100 3
+B	A	B-to-A_2	4	100 3
+B   D   B-to-D  1   100 4
+D   B   D-to-B  1   100 4
+B   D   B-to-D_3  1   100 6
+D   B   D-to-B_3  1   100 6
+B   D   B-to-D_4  1   100 7
+D   B   D-to-B_4  1   100 7
+B   D   B-to-D_2  4   100 5
+D   B   D-to-B_2  4   100 5
+
+NODES_TABLE
+name	lon	lat
+A	25	0
+B	31	-177
+C	60	-63
+D	62	37
+
+DEMANDS_TABLE
+source	dest	traffic	name
+A   D   100 dmd_a_d_1
