+++ pyNTM/model.py
@@ -12,6 +12,7 @@ to the latter's requirement to check for multiple Circuits between Nodes.
 
 from pprint import pprint
 
+import itertools
 import networkx as nx
 
 from .circuit import Circuit
@@ -245,6 +246,166 @@ class Model(MasterModel):
 
         self.validate_model()
 
+    def _route_demands(self, model):
+        """
+        Routes demands in input 'model'
+
+        :param model: input 'model' parameter object (may be different from self)
+        :return: model with routed demands
+        """
+
+        G = self._make_weighted_network_graph_mdg(include_failed_circuits=False)
+
+        for demand in model.demand_objects:
+            src = demand.source_node_object.name
+            dest = demand.dest_node_object.name
+
+            # Shortest path in networkx multidigraph
+            try:
+                nx_sp = list(nx.all_shortest_paths(G, src, dest, weight='cost'))
+            except nx.exception.NetworkXNoPath:
+                # There is no path, demand.path = 'Unrouted'
+                demand.path = 'Unrouted'
+                continue
+
+            # Find demand's load to interfaces it touches
+            num_paths = len(nx_sp)
+            demand_load_per_path = demand.traffic / num_paths
+
+            # all_paths is list of paths from source to destination; these paths
+            # may include paths that have multiple links between nodes
+            all_paths = []
+
+            for path in nx_sp:
+                current_hop = path[0]
+                this_path = []
+                for next_hop in path[1:]:
+                    this_hop = []
+                    values_source_hop = G[current_hop][next_hop].values()
+                    min_weight = min(d['cost'] for d in values_source_hop)
+                    ecmp_links = [interface_index for interface_index, interface_item in
+                                  G[current_hop][next_hop].items() if
+                                  interface_item['cost'] == min_weight]
+                    num_ecmp_links = len(ecmp_links)
+
+                    # Add Interface(s) to this_hop list and add traffic to Interfaces
+                    for link_index in ecmp_links:
+                        this_hop.append(G[current_hop][next_hop][link_index]['interface'])
+                    this_path.append(this_hop)
+                    current_hop = next_hop
+                all_paths.append(this_path)
+
+            path_list = self._normalize_multidigraph_paths(all_paths)
+            demand.path = path_list
+
+        self._update_interface_utilization()
+
+        return self
+
+    def _make_weighted_network_graph_mdg(self, include_failed_circuits=True, needed_bw=0, rsvp_required=False):
+        """
+        Returns a networkx weighted networkx multidigraph object from
+        the input Model object
+
+        :param include_failed_circuits: include interfaces from currently failed
+        circuits in the graph?
+        :param needed_bw: how much reservable_bandwidth is required?
+        :param rsvp_required: True|False; only consider rsvp_enabled interfaces?
+
+        :return: networkx multidigraph with edges that conform to the needed_bw and
+        rsvp_required parameters
+        """
+
+        G = nx.MultiDiGraph()
+
+        # Get all the edges that meet 'failed' and 'reservable_bw' criteria
+        if include_failed_circuits is False:
+            considered_interfaces = (interface for interface in self.interface_objects
+                                     if (interface.failed is False and
+                                         interface.reservable_bandwidth >= needed_bw))
+        elif include_failed_circuits is True:
+            considered_interfaces = (interface for interface in self.interface_objects
+                                     if interface.reservable_bandwidth >= needed_bw)
+
+        if rsvp_required is True:
+            edge_names = ((interface.node_object.name,
+                           interface.remote_node_object.name,
+                           {'cost': interface.cost, 'interface': interface, 'circuit_id': interface.circuit_id})
+                          for interface in considered_interfaces
+                          if interface.rsvp_enabled is True)
+        else:
+            edge_names = ((interface.node_object.name,
+                           interface.remote_node_object.name,
+                           {'cost': interface.cost, 'interface': interface, 'circuit_id': interface.circuit_id})
+                          for interface in considered_interfaces)
+
+        # Add edges to networkx DiGraph
+        G.add_edges_from(edge_names)
+
+        # Add all the nodes
+        node_name_iterator = (node.name for node in self.node_objects)
+        G.add_nodes_from(node_name_iterator)
+
+        return G
+
+    def _normalize_multidigraph_paths(self, path_info):  # TODO - static?
+        """
+        Takes the multidigraph_path_info and normalizes it to create all the
+        path combos that only have one link between each node.
+
+        :param path_info: List of of interface hops from a source
+        node to a destination node.  Each hop in the path
+        is a list of all the interfaces from the current node
+        to the next node.
+
+        path_info example from source node 'B' to destination node 'D':
+        [
+            [[Interface(name = 'B-to-D', cost = 20, capacity = 125, node_object = Node('B'),
+                    remote_node_object = Node('D'), circuit_id = '3')]], # there is 1 interface from B to D and a
+                    complete path
+            [[Interface(name = 'B-to-G_3', cost = 10, capacity = 100, node_object = Node('B'),
+                    remote_node_object = Node('G'), circuit_id = '28'),
+              Interface(name = 'B-to-G', cost = 10, capacity = 100, node_object = Node('B'),
+                    remote_node_object = Node('G'), circuit_id = '8'),
+              Interface(name = 'B-to-G_2', cost = 10, capacity = 100, node_object = Node('B'),
+                    remote_node_object = Node('G'), circuit_id = '18')], # there are 3 interfaces from B to G
+            [Interface(name = 'G-to-D', cost = 10, capacity = 100, node_object = Node('G'),
+                    remote_node_object = Node('D'), circuit_id = '9')]] # there is 1 int from G to D; end of 2nd path
+        ]
+
+
+
+        :return: List of lists.  Each component list is a list with a unique
+        Interface combination for the egress Interfaces from source to destination
+        example:
+
+            [
+                [Interface(name = 'B-to-D', cost = 20, capacity = 125, node_object = Node('B'),
+                    remote_node_object = Node('D'), circuit_id = '3')], # this is a path with one hop
+                [Interface(name = 'B-to-G_3', cost = 10, capacity = 100, node_object = Node('B'),
+                    remote_node_object = Node('G'), circuit_id = '28'),
+                 Interface(name = 'G-to-D', cost = 10, capacity = 100, node_object = Node('G'),
+                    remote_node_object = Node('D'), circuit_id = '9')], # this is a path with 2 hops
+                [Interface(name = 'B-to-G_2', cost = 10, capacity = 100, node_object = Node('B'),
+                    remote_node_object = Node('G'), circuit_id = '18'),
+                 Interface(name = 'G-to-D', cost = 10, capacity = 100, node_object = Node('G'),
+                    remote_node_object = Node('D'), circuit_id = '9')], # this is a path with 2 hops
+                [Interface(name = 'B-to-G', cost = 10, capacity = 100, node_object = Node('B'),
+                    remote_node_object = Node('G'), circuit_id = '8'),
+                 Interface(name = 'G-to-D', cost = 10, capacity = 100, node_object = Node('G'),
+                    remote_node_object = Node('D'), circuit_id = '9')]  # this is a path with 2 hops
+            ]
+        """
+        # List to hold unique path(s)
+        path_list = []
+
+        for path in path_info:
+            path = list(itertools.product(*path))
+            for path_option in path:
+                path_list.append(list(path_option))
+
+        return path_list
+
     def _make_circuits(self, return_exception=True, include_failed_circuits=True):
         """
         Matches interface objects into circuits and returns the circuits list
@@ -256,7 +417,7 @@ class Model(MasterModel):
                  comprised of two Interface objects
         """
 
-        G = self._make_weighted_network_graph(include_failed_circuits=include_failed_circuits)
+        G = self._make_weighted_network_graph_mdg(include_failed_circuits=include_failed_circuits)
 
         # Determine which interfaces pair up into good circuits in G
         paired_interfaces = ((local_node_name, remote_node_name, data) for
@@ -646,7 +807,7 @@ class Model(MasterModel):
         """
 
         # Define a networkx DiGraph to find the path
-        G = self._make_weighted_network_graph(include_failed_circuits=include_failed_circuits, needed_bw=needed_bw)
+        G = self._make_weighted_network_graph_mdg(include_failed_circuits=include_failed_circuits, needed_bw=needed_bw)
 
         # Define the Model-style path to be built
         converted_path = dict()
@@ -675,7 +836,7 @@ class Model(MasterModel):
         """
 
         # Define a networkx DiGraph to find the path
-        G = self._make_weighted_network_graph(include_failed_circuits=False, needed_bw=needed_bw)
+        G = self._make_weighted_network_graph_mdg(include_failed_circuits=False, needed_bw=needed_bw)
 
         # Define the Model-style path to be built
         converted_path = dict()
@@ -879,9 +1040,9 @@ class Model(MasterModel):
             pprint(interface)
             print()
 
-    def _make_weighted_network_graph(self, include_failed_circuits=True, needed_bw=0, rsvp_required=False):
+    def _make_weighted_network_graph_mdg(self, include_failed_circuits=True, needed_bw=0, rsvp_required=False):
         """
-        Returns a networkx weighted networkx digraph from
+        Returns a networkx weighted networkx multidigraph object from
         the input Model object
 
         :param include_failed_circuits: include interfaces from currently failed
@@ -889,11 +1050,11 @@ class Model(MasterModel):
         :param needed_bw: how much reservable_bandwidth is required?
         :param rsvp_required: True|False; only consider rsvp_enabled interfaces?
 
-        :return: networkx digraph with edges that conform to the needed_bw and
+        :return: networkx multidigraph with edges that conform to the needed_bw and
         rsvp_required parameters
         """
 
-        G = nx.DiGraph()
+        G = nx.MultiDiGraph()
 
         # Get all the edges that meet 'failed' and 'reservable_bw' criteria
         if include_failed_circuits is False:
@@ -906,16 +1067,18 @@ class Model(MasterModel):
 
         if rsvp_required is True:
             edge_names = ((interface.node_object.name,
-                           interface.remote_node_object.name, interface.cost)
+                           interface.remote_node_object.name,
+                           {'cost': interface.cost, 'interface': interface, 'circuit_id': interface.circuit_id})
                           for interface in considered_interfaces
                           if interface.rsvp_enabled is True)
         else:
             edge_names = ((interface.node_object.name,
-                           interface.remote_node_object.name, interface.cost)
+                           interface.remote_node_object.name,
+                           {'cost': interface.cost, 'interface': interface, 'circuit_id': interface.circuit_id})
                           for interface in considered_interfaces)
 
         # Add edges to networkx DiGraph
-        G.add_weighted_edges_from(edge_names, weight='cost')
+        G.add_edges_from(edge_names)
 
         # Add all the nodes
         node_name_iterator = (node.name for node in self.node_objects)
